+++ rasa_core/tracker_store.py
@@ -33,16 +33,18 @@ class TrackerStore(object):
         self.event_broker = event_broker
 
     @staticmethod
-    def find_tracker_store(domain, store=None):
+    def find_tracker_store(domain, store=None, event_broker=None):
         if store is None or store.store_type is None:
-            return InMemoryTrackerStore(domain)
+            return InMemoryTrackerStore(domain, event_broker=event_broker)
         elif store.store_type == 'redis':
             return RedisTrackerStore(domain=domain,
                                      host=store.url,
+                                     event_broker=event_broker,
                                      **store.kwargs)
         elif store.store_type == 'mongod':
             return MongoTrackerStore(domain=domain,
                                      host=store.url,
+                                     event_broker=event_broker,
                                      **store.kwargs)
         else:
             return TrackerStore.load_tracker_from_module_string(domain, store)