+++ CHANGELOG.rst
@@ -32,6 +32,7 @@ Added
 - http server (``rasa_core.run --enable-api``) evaluation endpoint
 - ability to add tracker_store using endpoints.yml
 - ability load custom tracker store modules using the endpoints.yml
+- ability to add an event broker using an endpoint configuration file
 
 Changed
 -------